+++ src/Client/Modals/ContextMenus/FooterTabs.fs
@@ -28,15 +28,15 @@ type FooterTabs =
                 |> Messages.InterfaceMsg
                 |> dispatch
 
-        let addDataMap =
-            fun _ ->
-                SpreadsheetInterface.UpdateDatamap(DataMap.init () |> Some)
-                |> Messages.InterfaceMsg
-                |> dispatch
+        // let addDataMap =
+        //     fun _ ->
+        //         SpreadsheetInterface.UpdateDatamap(DataMap.init () |> Some)
+        //         |> Messages.InterfaceMsg
+        //         |> dispatch
 
         let children (rmv: unit -> unit) : ReactElement seq = [
             Modals.ContextMenus.Base.Item("Add Table", addTable >> rmv, Icons.Table())
-            Modals.ContextMenus.Base.Item("Add Datamap", addDataMap >> rmv, Icons.Map())
+        // Modals.ContextMenus.Base.Item("Add Datamap", addDataMap >> rmv, Icons.Map())
         ]
 
         let mouseY = mouseY - 30